server/django_uv/pyproject.toml
[project]
name = "django-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.12",
    "django-cors-headers>=4.9.0",
    "django-stubs>=5.2.9",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
]
